Update patch set 1
Patch Set 1:

It took me some time to find out how to solve the issue and it's the first bigger change I submitted so be severe with me.

My first suggegstion was to do it like it was done in the PublishProductMojo, to create a ..plexus..StringSearchInterpolator within the RootPropertiesParser or the FeatureRootAdvice. But as I understood, the hole org.eclipse.tycho.p2.resolver.impl plug-in does not know anything about maven, plexus, maven session or the maven project and so I thought it might be a good idea to hide all this stuff and provide a component that could be injected that handles interpolation.

Patch-set: 1