Update patch set 1
Patch Set 1: Code-Review-1

I don't think we need the complexity of an interpolation OSGi service here.

As I mentioned earlier why don't we do interpolation on a lower (plain String value) level, as an implementation detail of BuildPropertiesParser for any value read from build.properties right away when parsing the file?

AFAIK when we introduced BuildPropertiesParser we made sure all parts of Tycho which need to access any values from build.properties use this interface, at least RootPropertiesParser is getting its information from BuildPropertiesParser. So the encapsulation/abstarcation for build.properties exists already, it should just be enhanced to replace maven variables in String values. we can simply hide ${variable} replacement behind the interface and apply it to all values (that is, if performance is good enough so we can always apply variable replacement on all values, regardless of whether they are used for root files or anything else that build.properties supports )

This is what I would expect as a user: If variable replacement is supported in build.properties, it should not matter on which value I'm using a variable.

Patch-set: 1
Label: Code-Review=-1