Update patch set 1
Patch Set 1:

Thanks Jan for the review and the comment. 
Ok, I see your point and I tried it the way, that the BuildPropertiesParserImpl does the work. 
I have issues with that: 
The actually parsing and reading the values from the build.properties file is done in the BuildPropertiesImpl and not in the BuildPropertiesParserImpl class. So my idea was to do interpolation in the BuildPropertiesImpl.splitAndTrimCommaSeparated and splitAndTrimValue Methods.
But to do the interpolation, the BuildPropertiesImpl class must know the maven session and maven project to get the properties from and since this class is within the org.eclipse.tycho.core.shared eclipse-plugin, this dependency to maven classes is not possible/wanted - right?

Is it ok to move the BuildPropertiesImpl from org.eclipse.tycho.core.shared to tycho-core next to BuildPropertiesParserImpl so that maven classes can be used?

Patch-set: 1