Update patch set 1
Patch Set 1:

Moving code around will probably introduce problems with the complicated separation of the "maven side" and the "embedded OSGi side" of Tycho code. These sides have only some of the *core bundles in common (and the *core bundles should not have any further dependencies).

I think it's better to do the  ${} replacement inside BuildPropertiesParserImpl. (it's a plexus component and you can inject the maven session via LegacySupport there as you did already for the InterpolationService if I'm getting it right)

I would propose to remove the BuildPropertiesImpl(File) constructor and only use BuildPropertiesImpl(java.util.Properties) instead (assuming ${} values in the properties are already replaced by BuildPropertiesParserImpl.

If we need ${} replacement in other places (not related to build.properties) too, we could even make it a plexus component ("InterPolationService" or similar), but I would do this only if there are other places that also need to replace ${} variables in Strings.

Patch-set: 1