442922 - Interpolate values in build.properties
Adding an interpolator that uses the plexus StringSearchInterpolator to
interpolate Strings. It hides all the plexus classes and does not expose
any of these. This interpolator is used in the BuildPropertiesParserImpl
after
parsing the build.properties file to interpolate each value. 
This interpolator is also used by the PublishProductMojo which used the
plexus one till now directly. 

Bug: 442922
Change-Id: I6ca3c6f4dc5874d338f0616a94d2e59c7c8f23ae
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>