453526 Removing unneeded testSuite parameter
The testSuite parameter had no effect, there were just  checks if
testSuite equals the current test module and if both, testSuite and
testClass are specified.
In a multi-bundle build, when the testSuite/testClass parameter is set
outside of the specific test bundle scope (e.g. by setting system a
system property or in a parent pom), the testSuite can now be removed,
but the "failIfNoTests" must be set to false.

Bug 453526
Change-Id: I612396b8a7a8c6ef5be0a4131f0d1b94849317d3
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>