Update patch set 5
Patch Set 5:

Sorry for spamming patchsets... the first 2 one I needed to figure out that it is not a good idea to implicit test the key order of a hashmap. 3rd an IT test failed, looked like an hudson env issue, with the 4rd I tried to trigger a new build and improved the test failure message, the 5th includes the fix for Jans comment - the osgi property for the default level is taken from the parameter in case it is specified there.

Patch-set: 5
Label: Verified=0