Update patch set 1
Abandoned

Thanks Igor for the explanation. In this case, I'd rather say that we document how to make the integration tests work - by calling one of the early phases in the Eclipse workspace.

Did you also see the error from the mirror goal disappear quickly in m2e, or was this just me?

Patch-set: 1
Status: abandoned