Update patch set 1
Abandoned

I actually made use of this permission recently to fix an accidental push, so IMHO we can keep this.

There are really a lot of clones around (e.g. on the CI server), so I don't really consider this a problem that committers have this permission.

Patch-set: 1
Status: abandoned