Update patch set 1
Patch Set 1:

> If we want to be really sure file does not get corrupted during write and rename, we can use java7 Files.move(ATOMIC_MOVE)

I'm not sure what we would gain with this over the current implementation.

Currently, there is the possibility that the index file is deleted and the rename no longer happens afterwards. In this case we'd lose the index, which isn't critical in most cases. With Java7 we could use java.nio.file.StandardCopyOption.REPLACE_EXISTING, which may solve this problem. But if we set that flag, we cannot use ATOMIC_MOVE because that flag causes all other ones to be ignored.

Patch-set: 1