Update patch set 7
Patch Set 7:

I've published a new version of the patch taking into account most of the comments. There are a couple things that I could not handle or had problems with and they are noted in the replies.

The biggest change from the initial commit is a change to the p2iu.xml file format. Indeed while reading the comment about the file format and the size attribute, I decided to simplify it even further to have it directly be the unit. The units top level node is no longer required.

Patch-set: 7
Label: Verified=0