Update patch set 10
Patch Set 10:

In this patch set, I've made the build order tests in P2IUTests harder by specifying the wrong order in the modules section. Now the assertions fail, but this doesn't matter because the tests implicitly use the deprecated LocalDependencyResolver. We'll need to make the tests work for bug 353889 though, so it is very good to already have them.

Patch-set: 10
Label: Verified=0