459873 Adding one additional level of temp directory for Equinox/P2
The default path for p2 stuff is at the same level as the
osgi.configuration.area. 
With this change, tycho will create a temp directory structure
${java.io.tempdir}/tychoXXXXequinox/config. This will be passed as
osgi.configuration.area property to Equinox. The p2 stuff will then end
up in ${java.io.tempdir}/tychoXXXequinox/p2 and will so be cleaned up
after the build because the whole temp directory "tychoXXXequinox" will
be deleted.
 
Bug: 459873
Change-Id: I9edadd53dbddffa576faab80b40ed6328eb8d18c
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>