Update patch set 12
Patch Set 12:

(11 comments)

All addressable concerns have been taken care of.
The duplicate RepositoryWithIUTest was a left over from the time where the fix for the handling of iu tag in category.xml was included in this change set. I have removed this tag from the change set since it is included in the other one.

Patch-set: 12
Label: Verified=0