460487 Support compiler logs for multi-jar bundles
If the compiler argument -log is specified in the pom, and the bundle is
configured to produce more than one jar, pass different log file paths
for the different jars to the compiler. The new log file name is
computed with <path of configured log file>/<library path>_<library
name>_<configured log file name> (e.g.
/path/to/log/libFolder_library.jar_logfile.xml)

Bug: 460487
Change-Id: I6f92e1ec13cc8889127b108ead5f9bc7f4366241
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>