460487 Support compiler logs for multi-jar bundles
Adding 2 parameters that control if and how the compiler output should
be persisted. <logDirectory> specifies the path to the log files, <log>
the format ("plain"/"xml"). If omitted, compiler logs will not be stored
into files. The filename for "main" classes is "@dot", for other jars
the filename is derived from the jar name. 
If -log is present and <log> is set, the values of the
new 2 parameters overwrite the -log specified in compiler arguments.

Bug: 460487 
Change-Id: I6f92e1ec13cc8889127b108ead5f9bc7f4366241 
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>