Update patch set 1
Patch Set 1:

I wonder if the dependency to https://git.eclipse.org/r/#/c/41895/2 is a hard one.

If we only need the adoption of changes in the p2 API (and not other pending fixes/TODOs for bug 428889), maybe we can split up https://git.eclipse.org/r/#/c/41895/2 into one change for bug 428889 and one change that only adapts to the changes in p2 so this one can depend on the latter and thus go in earlier?

Patch-set: 1