Update patch set 1
Patch Set 1:

Just to be clear as to why I'm suggesting to revert the p2 changes. This is simply because progress need to be made toward the mac layout app and I'm trying to limit the changes that we have to do (you can also call that dependencies or obstacle) and removing the p2 changes in the publisher would then leave you all the time you need to complete the code changes.
Once everything will be good on your side, then the publisher cahnges can be reintroduced and everything will be rosy.

Patch-set: 1