461651 No default value for compiler plugin parameter 'log'
The 'log' compiler plugin parameter must not have a default, since if
present, logging into a file is enabled. If logging is enabled with that
new parameter and a custom compiler argument '-log' is specified, the
build will now fail with an error message that informs the user that
only either of them ('log' plugin parameter or '-log' compiler argument)
is allowed. 

Bug: 461651
Change-Id: I7b7c0e56f3de839c2980f4949c259e66c678b0ad
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>