Update patch set 2
Patch Set 2: Code-Review+1

> What do you think about Davids comment in the bug, that failing the build in that case might not be the best solution? If we change that to just log a warning, we do not need to care about an exception here.

IMHO, it is important that the user get a proper message, and not the cryptic error that the user currently gets. This is implemented with this change, so we should submit it. (Not sure if you still wanted to do something with the test setup though.)

I agree that not failing but just warning is nicer, but this is hard to get correctly, so if done at all, this should be a separate commit.

Patch-set: 2
Label: Code-Review=+1