461586 Enable log output when fetching artifacts.jar files
Firstly, RepositoryLoader now has an extra progressMonitor field, which is used
when calling repositoryManager.loadRepository. Setting this to null makes
RepositoryLoader work as before (FileRepositoryArtifactProviderTest has been
changed to do this).

Secondly, RepositoryArtifactProvider now has a 4-arg constructor that takes a
progress monitor; the progress monitor is used when creating a RepositoryLoader.
The existing 3-arg constructor delegates to the new 4-arg constructor, passing
null, for backwards compatibility.

Finally, TargetPlatformFactoryImpl.createRemoteArtifactProvider now wraps the
MavenLogger with a DuplicateFilteringLoggingProgressMonitor, and passes that
progress monitor when creating a RepositoryArtifactProvider.

Bug: 461586
Change-Id: I2bcb3f02800c18e1a450f413b1d87f4cca38c49f
Signed-off-by: Richard Fearn <richardfearn@gmail.com>