Update patch set 1
Patch Set 1:

> (1 comment)

Ah, thanks! In the initial patch on bug 461586, I constructed the progress monitor inside RepositoryLoader. But over the weekend I decided it was better to create it once, in TargetPlatformFactoryImpl, and pass it down to the RepositoryLoader (through the RepositoryArtifactProvider).

I'll upload a new patch set later.

Patch-set: 1
Label: Verified=0