Update patch set 1
Patch Set 1: Code-Review-1

> Personally, I prefer to keep reference to bugzilla bug id test
 > names. It gives context of the usecase(s) the test is supposed to
 > cover, which I find very useful. "baseline" does not really say
 > much unless you read https://bugs.eclipse.org/bugs/show_bug.cgi?id=362883.

So then maybe we should call the project "reproducibleArtifacts"?

I forgot to keep a reference to the bug number in the test class and pom.xml. This still needs to be fixed. The pom.xml and/or test class is the right place to put the bug number, because then one IT can cover multiple bugs.

BTW, the test class doesn't have a copyright header. Igor, could you recover it?

Patch-set: 1
Label: Code-Review=-1