Update patch set 3
Patch Set 3: Code-Review-1

Hm... The problem is that this change doesn't solve bug 463670. So we need to mark the switch as experimental so that we can remove it again in case it doesn't make any difference.

Patch-set: 3
Reviewer: Gerrit User 489 <489@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1