Update patch set 1
Patch Set 1:

> maybe put it in a "setup" subfolder and add a README.md next to it
 > describing a) how to use it (for contributors) and b) how to change
 > it/author it (for committers)
 > 
 > the top-level README.md could reference this README.md
 > 
 > did not try it out yet but one question I had is how does it
 > reference the existing tycho-bundles.target file which we have
 > checked in?
 > 
 > Before we can publish it to the catalog we need some testing by
 > several people to make sure it works in different environments.

I will move it and will create a README.md with detail infos how to use it for now (without having it in the Oomph Product Catalog) and how to change it. 

The target platform is set as one of the last setup tasks. First it clones the Tycho git repo and then unsing the target definition file from tycho-bundles/tycho-bundles-target/tycho-bundles-target.target as target platform

Patch-set: 1