Update patch set 3
Patch Set 3:

(1 comment)

Have you double clicked on the Tycho setup file on the 2nd page, so that Tycho is shown in the bottom table of things that get installed? Thats a little bit tricky in the Installer UI, that selecting the Install file alone is not enougth, you must doubleclick it.

Patch-set: 3