Update patch set 3
Patch Set 3: -Code-Review

thanks for your help Martin I was finally able to test the setup file and it worked for me.

BTW the few remaining initial build errors in the tycho workspace are probably caused by bug 464491.

Martin, please don't take this against your efforts, they are very much appreciated but sorry, the usability of the installer IMHO right now is quite bad.

E.g. why do I have to drag and drop the setup file (just adding it using the "+" icon won't work) and it has to be into a very specific folder? No hints at all for this hidden feature.

why double-click on the user setup folder otherwise simply using "Next" won't work?

If we want to make it easier for users, this is more complicated than it should be and kind of defies the original point to make the project setup easier.

I hope that once we have it in the oomph catalogue, end user experience will be much smoother (just guide the user through the wizard and "Next" will always do the right thing).

Maybe I'm getting it wrong and cryptic usability is normal and expected in "expert mode". As long as it will be easy for end users in the end, I'm fine.

I think all crucial hints Martin gave in the gerrit comments should be added to the README file at least for the committer/development of the setup file  workflow.

Patch-set: 3