Update patch set 3
Patch Set 3:

Jan I also think that the setup process using the setup file does have some sticking points. Onces the setup file is in the catalog the users do not have to do the drag and drop but the double click still needs to be done. I dont have any problems with not using the Ooomph installer for now. But I think then we should give feedback to the Oomph Team so that they know where we see problems using the installer.

Patch-set: 3