Update patch set 3
Patch Set 3:

no problem. tearing apart the commit into two commits after the fact is probably not worth the effort. I will review it this way now. Just a hint for next time to make it easier to read for the reviewer. Give me some time to dig through this, I'm targeting end of this week.

Patch-set: 3