[WIP] Allow contributing system packages.
It's possible to contribute additional bundles to the Tycho
Eclipse runtime. What's lacking is the ability to add extra
system packages. Those extra packages are intended to share 
code between the OSGi environment and the Maven runtime.

This is a work in progress. What's missing is a way to actually 
make the packages available to the class loader loading the 
Equinox framework. Only listing the packages isn't enough.

Change-Id: I28506c5dd9d0fd64b5be6c19f6a1362b3d64b817
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>