Bug 463670: correctly handle literal ".qualifier"
Detect Luna or earlier executable feature by comparing against
the Luna SR2 version instead of Mars SR0.
This way, a Mars version even with literal ".qualifier" will
be correctly detected as Mars or later as the significant difference
is already in the micro (third) digit (102 in Luna vs 200 in Mars).

Change-Id: I79d177444737dc2bb4e69ff5c789184dd0dc5b30
Signed-off-by: Jan Sievers <jan.sievers@sap.com>