Bug 474885 Do throw a ProvisionException with repo not found code
In case of an offline build, when a specific file in the cache could not
be found (e.g. the cached content.xml) we throw a ProvisionExcpetion.
Only in case of an error code ProvisionException.REPOSITORY_NOT_FOUND,
p2 does try to download the artifact with other prefixes (e.g.
compositeContent.xml). This change does set the code, so that p2 does
not fail e.g. if the cached content.xml could not be found but proceed
trying it with compositeContent.xml, content.xml.xz. This is only done
in offline mode where we do not know what type of p2 repo we have.
Extending the OfflineModeTest to also test with a composite and a XZ
repository.

Bug: 474885
Bug: 461787
Change-Id: I857696fd011f50faf2777576962fbb17d89117a4
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>