Bug 474885 Do throw a ProvisionException with repo not found code
In case of an offline build, when a specific file in the cache could not
be found (e.g. the cached content.xml) we throw a ProvisionExcpetion.
Before this change, we did that without specifying a code in the error
status. In such a case p2 fails fast and does not try to download (in
offline mode get from cache) other files.
In offline mode we do not know if either content.xml,
compositeContent.xml or content.xml.xz are in the cache so p2 must not
fail fast but should try all of the possible files. P2 does that if the
ProvisionException contains the error code
ProvisionException.REPOSITORY_NOT_FOUND. This change does set this error
code in the status of the ProvisionException and so p2 does not fail
e.g. if the cached content.xml could not be found but proceed trying it
with compositeContent.xml, content.xml.xz. 
Extending the OfflineModeTest to also test with a composite and a XZ
repository.

Bug: 474885
Bug: 461787
Change-Id: I857696fd011f50faf2777576962fbb17d89117a4
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>