Bug 357513: add MD5 checksums to artifact repo
force re-creation of artifact repo so that
MD5 checksums are added.

creating them only when the repo is aggregated
(as opposed to when the bundle jar is created) has the advantage
that there can be no checksum-changing operations on individual
artifacts like jar signing or pack200.
p2's RecreateRepositoryApplication will calculate and add

"download.md5" and "artifact.size"

properties to all artifacts aggregated in target/repository.

Added corresponding IT.
Verified manually using tycho demo project
http://git.eclipse.org/c/tycho/org.eclipse.tycho-demo.git/tree/itp04-rcp
that the only effect of RecreateRepositoryApplication
is to add above 2 properties to artifacts.xml.

Change-Id: Idfb95af01f63679998d4295331b98096aae88ed7
Signed-off-by: Jan Sievers <jan.sievers@sap.com>