Bug 480951 Ignore unrelated submodules when checking git dirty state
_Setting the sub-module-mode flag to IgnoreSubmoduleMode.ALL in order to
filter out the git state of unrelated submodules when checking the diff
of a git repo. If the submodule is part of the build, the diff still
contains it changes, because the path filter does not filter it out.
_Added an IT "dirtySubmodules" that checks that the build fails if there
are dirty submodules that are part of the build
_Added an IT "dirtyUnrelatedSubmodules" that checks that the build does
not fail if there are dirty submodules that are not realted to the
build.

Bug: 480951
Change-Id: Ib1b219927035aaaa5debecebd66eea575a9a9cde
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>