Update patch set 5
Patch Set 5:

Thanks Jan for the hints regarding the TestNG bundle stuff.
Not sure if I got everything correctly. The PatchSet 5 contains such a "compile time bundle" that pulls in the TestNG classes in order to make the Tycho build work. However I had not use the "unpack" instead of the "copy" goal because the TestNG CQ says that I have to remove the jquery.js which is inside the original TestNG jar. 

Regarding the ITs I did made kind of the same bundle as you have seen, but I can change that to your suggested solution.

Patch-set: 5
Label: Verified=0