Update patch set 6
Patch Set 6:

yes going with option 2 is fine.

for the ITs, you are right the benefit would be we can easily add new tests for the testNG scenario. If it's not needed right now, maybe I was thinking too far ahead and we can just keep it as is though.

Patch-set: 6
Label: Verified=0