Update patch set 7
Patch Set 7:

I removed the "Work in progress" since IMHO its now in a state where it can be reviewed. 
The open thing is that if that test provider should be enabled, it has to be done with the providerHint tag in the test configuration. But I think that's ok for now. I will create a request that the TestNG jar get's part of the Orbit project and onces this is done and we have a TestNG OSGi bundle, we can change it to search the class test class path.

Patch-set: 7