Update patch set 2
Patch Set 2:

does this fix the failing build of the sample project in bug 407273 ?
if so, can you add a stripped-down version of it as an IT?

From the commit message it's not very clear if this depends on another fix in p2 to actually work. If so, which p2 bug does it depend on?

Patch-set: 2
Label: Verified=0