482487 Using the first .product file found to extract model info
If the project is neither a plugin/test-plugin nor a feature, it looks
if there is a .product file in the directory. 
If so, it uses the "uid" attribute of the root element as artifact id
and the "version" attribute as artifact version. 
It adds the 2 goals ("archive-prodcuts" and "materialize-products").
If there are multiple .product files in the directory, the first one
that was found is used.

Bug: 482487
Change-Id: I185d99185aba52e58cbf82a8e850c0448498b305
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>