501809 Adding a tycho specific MirrorSelector to support URL prefixing
The MirrorSelector component is used to select a mirror for a given
ArtifactRepository. The TychoMirrorSelector does compare the artifact
repository URL with the "mirrorOf" definition of the mirrors. If the
repo URL starts with "mirrorOf" a new Mirror object is created with the
URL containing the mirror URL with the remaining path of the repository
URL. The "old" way (repo "id" matches mirrors "mirrorOf" attribute)
still works and if there is a mirror defined that way, this mirror is
used.

Bug: 501809
Change-Id: I4718fc22f76d782c452957df706a8c87c0c8c119
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>