367945 Using the maven.build.timestamp which is the session start time
Maven propagates the session start time as maven.build.timestamp. The
DefaultBuildTimestampProvider does return that start time instead of
building its own timestamp and keep it in the user properties.
The added IT test builds a project with a maven project and 2 eclipse
plugins. The maven.build.timestamp is written (using resource filtering)
into a text file and the test compares the test file content to match
the eclipse plugins Bundle-Version of the manifest file.

Bug: 367945
Change-Id: I8ee7bc7d9ae7708ede2a0a2b71ecc1e90eff3779
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>