Update patch set 1
Patch Set 1: Code-Review-1

As said, thats a quick and dirty Solution. It checks if there is the .polyglot.build.properties "pom" file available and uses that. When writing only "real" pom files are written, not the polyglot ones.

I tried to use the PolyglotModelManager to get a ModelReader for a specific project and store the Model and the MutablePom (for existing pom files) in the projects metadata and use the model to get the information about version, packaging etc... 
But to get that work, we must add the tycho-extras pomless artefact as dependency to this project in order to get the Tycho pomless classes injected and I think that we must not have depenencies from tycho to tycho-extras right?

Patch-set: 1
Label: Code-Review=-1