Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > without having looked into the details, my high level assumption
 > would be that we have to check (maybe in several places in
 > tycho-versions-plugin) if the pom.xml file exists and if not,
 > simply ignore it as we don't have to do any pom.xml changes in case
 > of pomless.

That's right but currently the information read from the pom (stored in the MutablePomFile instance) is used for many manipulators not only for the manipluator that changes the pom file. The maven coordinates are taken from the MutablePomFile instance in many places to do checks. 
I thought about using for such things the MavenProject but in some cases, we are changing stuff in projects that are not in the reactor. E.g. when a module is part of an inactive profile. Currently such bundles would get updated even if they are not part of the reactor.

Patch-set: 1