Update patch set 2
Patch Set 2:

> not sure if we should fail the build if no .product file found
 > 
 > it's OK to have just a category.xml in packaging type
 > eclipse-repository
 > 
 > in case of only category.xml, we don't have a version and just skip
 > (maybe add a warning)

Oh thanks - absolutly right for eclipse update sites (eclipse-repository packaging and a category.xml in the directory)!

Now I log a warning. I have been torn between log a warning or just an info, because it is a valid usecase that there is no product file but silenty skipping might not be useful. I went with a warning but can change it to an info if you think that might be better.

Patch-set: 2
Label: Verified=0