Bug 495353: Update surefire plugins to v2.19.1
Updating the surefire plugins to the most recent version (v2.19.1).
This update includes 2 additional parameters that could be passed to
surefire (and could be configured in the tycho test mojo):
_rerunFailingTestsCount
_skipAfterFailureCount

The changes in the TestMojoTest are required because the directory
scanner api changed and it seems, that this scanner does append .class
at the end of include/exclude patterns (if not already specified), so it
was not possible anymore to use .txt files for testing the scanForTests
method. Because of that, I changed the test that the test method will
create 2 test dummy .class files do to the scanForTests.

CQ: 12887
CQ: 12893
CQ: 12886
CQ: 12892
CQ: 12891
CQ: 12896
CQ: 12889
CQ: 12895
CQ: 12894
CQ: 12888
CQ: 12890
Bug: 495353
Change-Id: I838b69f77504c9641ccb80d7e11a3bcda8032a62
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>
Also-by: Jan Sievers <jan.sievers@sap.com>