Bug 519280: Fix issue when dots are used in the <folder> section
When using dots in the <folder> section of a root.folder property Tycho
shows either an error message (when the path starts with dots) or
silently ignores the property line (when dots are used within the folder
name). This fix solves this problem and provides a JUnit test case to
verify the fix is there.

Change-Id: Ic5e8577f02a7c2aa38ff8a1fae1770759072b510
Signed-off-by: Franz Polzer <franz_polzer@mentor.com>