Update patch set 1
Patch Set 1:

> I would double-check original bug
 > 
 > https://bugs.eclipse.org/bugs/show_bug.cgi?id=368985
 > 
 > which has some links about how license features work in PDE.
 > 
 > Also check  corresponding test
 > https://github.com/eclipse/tycho/blob/master/tycho-its/src/test/java/org/eclipse/tycho/test/licenseFeature/LicenseFeatureTest.java
 > 
 > which uses sample project
 > https://github.com/eclipse/tycho/tree/master/tycho-its/projects/licenseFeature
 > 
 > the new test should probably be added there.
 > 
 > not sure but if I get it right the problem only happens if license
 > and license URL are not extracted to feature.properties and
 > referenced via something like %license and %licenseURL ?

Thanks Jan,
Till now Tycho does only merge the feature.properties file (copy the content of the license feature features.properties into the feature.properties of the regular feature. 
If you would like to have the license URL and license text taken from that properties file, your regular feature's feature.xml had to contain e.g. : 

<license url="%licenseURL">%license</license>

PDE does the following:
_Copy the content from the feature.properties of the license feature to the feature.properties of the regular feature
_Set the values from the license feature's 

 > I would double-check original bug
 > 
 > https://bugs.eclipse.org/bugs/show_bug.cgi?id=368985
 > 
 > which has some links about how license features work in PDE.
 > 
 > Also check  corresponding test
 > https://github.com/eclipse/tycho/blob/master/tycho-its/src/test/java/org/eclipse/tycho/test/licenseFeature/LicenseFeatureTest.java
 > 
 > which uses sample project
 > https://github.com/eclipse/tycho/tree/master/tycho-its/projects/licenseFeature
 > 
 > the new test should probably be added there.
 > 
 > not sure but if I get it right the problem only happens if license
 > and license URL are not extracted to feature.properties and
 > referenced via something like %license and %licenseURL ?

Thanks Jan,

Well IMHO it is not only the %license and %licenseURL thing. Currently the shared license feature with Tycho is only working if you e.g. have:

<license url="%licenseURL">%license</license>

within the feature.xml of the feature that uses the shared license feature. IMHO that's an issue, because if one changes the key in the feature.properties file (of the shared license feature), all feature.xml files that uses that share license feature needs to be updated.

If I see it correctly PDE does
_Copy the content of the feature.properties from the license feature into the regular feature one.
_Replace the values of <license> within the regular's feature.xml with the one from the shared license feature.

With that (PDE) behaviour, not only the %license thing would work, but also if one is specifing fixed values in the shared license feature's feature.xml

This fix addresses replacing the <license> values in the feature.xml. There is another issue (See comment #4 in the bug), that the regular feature needs to have a feature.properties. If that file is missing (which is a valid case) in the regular feature, Tycho does not copy the feature.properties from the license feature.

I will prepare another patchset with the 2nd issue and will add an itest that shows tests the new (additional) behavouir.

Patch-set: 1
Label: Verified=0