Bug 519941 Copy the license features's feature.properties
As before the feature.properties in the target directory is the content
of the feature's feature.properties and the content of the license
feature's feature.properties (appended to it if there is a license
feature and it does have a feature.properties).
The new behavior is, that if the local feature does NOT have a
feature.properties but the license feature does, the license feature's
properties file will be copied.

Change-Id: I0bb49a6563a6ac181b9049c8d8a61d165a7fff96
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>