Update patch set 1
Patch Set 1:

it looks like we will first have to upload the new eclipse binaries as SNAPSHOTs with

https://hudson.eclipse.org/tycho/job/tycho-deploy-dependencies/

also, can you please open a bug or comment on existing bug

https://bugs.eclipse.org/bugs/show_bug.cgi?id=514471

if appropriate so we have more information on what exactly we are working on here. e.g. what's the milestone version of JDT we need?

we will also need an (integration) test to prove the new functionality works.

Patch-set: 1
Label: Verified=0