Update patch set 6
Patch Set 6:

Martin, I believe you were close but there was an Import-Package missing which I believe is the cause for most of the IT failures. I uploaded a new patchset now which should improve the test failure rate, let's wait for the result.
I think this is an important prerequsite for being able to add JUnit 5 support

Patch-set: 6
Label: Code-Review=0