Update patch set 1
Patch Set 1:

(2 comments)

I like the approach.

As of now I think it effectively changes the default from JavaSE1.6 to JavaSE-1.7 if I read the global default https://wiki.eclipse.org/Tycho/Execution_Environments right.

IIRC Tobias introduced various places in Tycho code that specifically handle the special a.jre.javase 1.6  IU, so there is a certain risk to change the default here. I'm also unsure about implications for products being installed from repos Tycho produces as they have a perfect version match dependency on this IU.
I think it should be an explicit opt-in for the user and the default should stay as is for now simply to reduce risk. 
I'm not sure about reusing the EE of the target-platform plugin. Most people never specify it and rely on bundle BREE instead. The EE used by the build is something different than an EE published as a side-effect of publishing a product. Rather make it a mojo parameter of PublishProductMojo (with default value JavaSE-1.6)
We have the Tycho 1.1 release immediately pending for which I would like to minimize risk of regressions (I think you've seen with the Java 9 support how hysteric eclipse platform is about any incompatible changes and you never know what detail they rely on). If we put this into Tycho 1.2, we can probably change defaults and announce it as we have enough time to test if there are  the ripples caused for downstream builds/p2 installs.
If you want this to be in Tycho 1.1 (I would prefer to wait though), it should not change the default value.

Patch-set: 1