Update patch set 1
Patch Set 1:

(1 comment)

Yes, I think a mojo parameter is less risky and more powerful.
There is a flaw in how product defines the compatibility, and as suggested in the bug, it's more an issue in p2 defining a requirement on the IU instead of a require capability.

Patch-set: 1
Reviewer: Gerrit User 587 <587@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0