Bug 530207 - Update the empty *a.jre* IUs to Java 9
StandardEEResolutionHints adds empty IUs in order to be able to resolve
org.eclipse.sdk.ide but these IUs are at version 1.6. In Bug 530093 and
Bug 530150   Eclipse moves these to Java 9. Tycho has to adjust it's
workaround otherwise trying to using tycho with Photon target ends up
with:
[ERROR] Cannot resolve target definition:
[ERROR]   Software being installed: org.eclipse.platform.ide
4.8.0.I20180123-1010
[ERROR]   Missing requirement: org.eclipse.platform.ide
4.8.0.I20180123-1010 requires 'config.a.jre.javase [9.0.0]' but it could
not be found

Change-Id: Iac77086de0a345713b4a73fd10edf647131b829f
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>