Update patch set 10
Patch Set 10:

> @Martin it looks like there are some CQs missing?
 > 
 > there should be a new helper script
 > 
 > ./tycho-releng/helpers/list_unapproved_dependencies.sh
 > 
 > which lists the unapproved dependencies to be added to
 > 
 > ./tycho-releng/helpers/approved_dependencies.txt

Thanks Jan, your right, a couple of CQs were missed. I created them. 

The helper script seems not to work for me. It still shows many dependencies as missing, but they are - in the right version - in the approved_dependencies list...? But anyway, once the new CQs are approved, we should be fine.

Patch-set: 10