Update patch set 11
Patch Set 11:

> Because we can still resolve against old p2 repos that do not have the osgi.ee provided capabilities(as the build of the sample project using Mars proves)

When running integration tests, I think it's not necessary to change existing repos because some of the available mechanisms are here to provide the JRE units anyway.
But for unit tests, those mechanisms are not covered nor used and it has been necessary to complete the JRE ius to get resolution working in "unit-er" tests.

Patch-set: 11
Label: Verified=0